詳細画面に参加しているイベント一覧追加　デフォルト画像表示

+++ persol_users/views.py
@@ -30,7 +30,7 @@ def user_add_operation(request):
                 data_tmp = request.FILES['data']
                 
             except:
-                data_tmp = ''
+                data_tmp = 'user_image/default.PNG'
             
             finally:
                 q = PersolUser(
@@ -76,14 +76,15 @@ def user_modify(request):
                 data_tmp = request.FILES['data']
                 
             except:
-                data_tmp = ''
+                data_tmp = 'user_image/default.PNG'
             
             finally:
                 user.data = data_tmp
                 user.save()
                 
                 if tmp:
-                    os.remove(tmp)
+                    if tmp != BASE_DIR + '/user_image/default.PNG':
+                        os.remove(tmp)
                     
                 return HttpResponseRedirect(reverse('persol_users:index'))
